+++ lib/details/models/artist_model.dart
@@ -8,7 +8,7 @@ class ArtistModel {
   Map<dynamic, dynamic> _info;
 
   /// Return artist [id]
-  int get id => int.parse(_info["_id"]);
+  int get id => _info["_id"] is String ? int.parse(_info["_id"]) : _info["_id"];
 
   /// Return artist [artistName]
   String get artistName => _info["artist"];
@@ -17,15 +17,20 @@ class ArtistModel {
   // String get artistKey => _info["artist_key"];
 
   /// Return artist [numberOfAlbums]
-  String get numberOfAlbums => _info["number_of_albums"];
+  int? get numberOfAlbums => _info["number_of_albums"] is String
+      ? int.parse(_info["number_of_albums"])
+      : _info["number_of_albums"];
 
   /// Return artist [numberOfTracks]
-  String? get numberOfTracks => _info["number_of_tracks"];
+  int? get numberOfTracks => _info["number_of_tracks"] is String
+      ? int.parse(_info["number_of_tracks"])
+      : _info["number_of_tracks"];
 
   /// Mp3 only support one image type, so, Artist Image don't exist.
   ///
   /// To get Artist Image it's necessary to use an external database.
-  String get artwork => "";
+  @Deprecated("message")
+  Uint8List? get artwork => _info["artwork"];
 
   // ///Return artist [artwork]
   // String get artwork => _info["artwork"];